remove use client from pages
