+++ client/src/app/about/page.tsx
@@ -1,7 +1,10 @@
-'use client';
-import { type NextPage } from 'next';
+import { Metadata, type NextPage } from 'next';
 
 import Wrapper from 'containers/wrapper';
+
+export const metadata: Metadata = {
+  title: 'TNC Prototype Dashboard',
+};
 
 const About: NextPage = () => {
   return (
