+++ client/src/app/projects/[id]/page.tsx
@@ -1,6 +1,10 @@
-import { type NextPage } from 'next';
+import { Metadata, type NextPage } from 'next';
 
 import ProjectDetail from 'containers/projects/detail';
+
+export const metadata: Metadata = {
+  title: 'TNC Prototype Dashboard',
+};
 
 const Project: NextPage = () => {
   return <ProjectDetail />;
