+++ client/src/app/resources/page.tsx
@@ -1,7 +1,10 @@
-'use client';
-import { type NextPage } from 'next';
+import { Metadata, type NextPage } from 'next';
 
 import Hero from 'containers/home/hero';
+
+export const metadata: Metadata = {
+  title: 'TNC Prototype Dashboard',
+};
 
 const Resources: NextPage = () => {
   return <Hero />;
