--- client/src/containers/meta-tags/types.d.ts
@@ -1,15 +1,0 @@
-export interface MetaTagsProps {
-  name: string;
-  title: string;
-  description: string;
-  url?: string;
-  type: string;
-  publisher?: string;
-  section?: string;
-  tag?: string;
-  imageURL?: string;
-  imageWidth?: string;
-  imageHeight?: string;
-  twitterCard?: string;
-  twitterSite?: string;
-}
